Update path references
Move all branch related files into completely separate branch dirs instead of having multiple dirs
with their own branch subdirs. Move log out of "source" dir (formerly "active" dir) which now
is stricture for holding source code (data from github) including temporarily saved previous
versions of repos. Add backup dir to branch dir and add init of branch and log dir to builder.
Simplify all util path references to be relative to files instead of relying on user home
directory.

+++ tools/builder.sh
@@ -26,38 +26,58 @@ echo -e "\n"
 
 
 # setup branch directory
-src="${HOME}"/geo/"$branch"
+branch_dir="/sciclone/aiddata10/geo/${branch}"
+src="${branch_dir}/source"
 
-# rm -rf "$src"
-find "$src" -type f -exec rm -rf "{}" \;
-find "$src" -type d -exec rm -rf "{}" \;
+if [[ -d ${src} ]] && [[ $2 == "--overwrite" ]]; then
+    find "$src" -type f -exec rm -rf "{}" \;
+    find "$src" -type d -exec rm -rf "{}" \;
+fi
 
-mkdir -p "$src"/{tmp,git,latest,log}
-#,'jobs',tasks}
+if [[ -d ${src} ]];then
+    echo "Warning: Source directory exists for branch (${branch_dir})"
+    echo "    Use --overwrite option to rebuild source for branch"
+    echo "    (Other directories in branch will not be removed)"
+fi
 
-cd "$src"
+
+mkdir -p "$src"/{git,latest}
 
 
-# clone tmp geo-hpc for init scripts
-git clone -b "$branch" https://github.com/itpir/geo-hpc tmp/geo-hpc
-ln -sfn tmp/geo-hpc "$src"/geo-hpc
+# create temp dir, clone geo-hpc for init scripts
+tmp=$(mktemp -d)
+git clone -b "$branch" https://github.com/itpir/geo-hpc "$tmp"/geo-hpc
+
 
 # run load_repos.sh
-bash "$src"/tmp/geo-hpc/tools/load_repos.sh "$branch" #2>&1 | tee "$src"/log/load_repos/$(date +%s).load_repos.log
-
+bash "$tmp"/geo-hpc/tools/load_repos.sh "$branch" #2>&1 | tee "$src"/log/load_repos/$(date +%s).load_repos.log
 
 # setup crons using manage_crons.sh script
-bash "$src"/tmp/geo-hpc/tools/manage_crons.sh "$branch" init
-
+bash "$tmp"/geo-hpc/tools/manage_crons.sh "$branch" init
 
 
 # add basic config info to mongo (for det, etc.)
 # does not update when config changes
 #   may change later, currently just needs basic branch name/server info
-python "$src"/tmp/geo-hpc/tools/add_config_to_mongo.py "$branch"
-
+python "$tmp"/geo-hpc/tools/add_config_to_mongo.py "$branch"
 
 
 # clean up tmp geo-hpc
-rm -rf "$src"/tmp/geo-hpc
+rm -rf "$tmp"
 
+
+
+# initialize other directories in case they do not exist
+
+mkdir -p "${backup_dir"/log
+
+backup_dir="${branch_dir}"/backups
+mkdir -p "${backup_dir}"/{mongodb_backups,mongodb_logs}
+
+
+output_dir="${branch_dir}"/outputs
+mkdir -p "${backup_dir}"/{det,extracts,msr}
+
+
+find "${branch_dir}" -type d -exec chmod u=rwx,g=rwxs,o=rx {} +
+find "${branch_dir}" -type f -exec chmod u=rw,g=rw,o=r {} +
